requirements.txt
Qt6-tools>=6.5
PyQt6>=6.6
PySide6>=6.6
vtk>=9.3
itk-tubetk>=1.3.5
